+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 ### Changed
 
 - `all-is-cubes` library:
+    - `block::EvaluatedBlock` now has a `voxels` field of the new type `Evoxels`, which replaces the previous `resolution` and `voxels` fields.
+      This simplifies the data model, in that there is now _always_ a set of `Evoxel`s defining a block's shape, even if there's only one of them, and it is always found in the `voxels` field. This is a breaking change for code that accesses `EvaluatedBlock` data.
     - `block::AIR` now has its own dedicated primitive, `Primitive::Air`. The behavior is unchanged.
     - `universe::UniverseTransaction::insert()` now takes a `URef` created by `URef::new_pending()`, instead of a bare value.
       This allows associations between the new member and other objects to be created within the same transaction.